requirements.txt
asgiref==3.5.2
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.1.3
django-cors-headers==3.14.0
django-extensions==3.2.1
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-jsonapi==6.0.0
huepy==1.2.1
idna==3.4
inflection==0.5.1
Jinja2==3.1.2
livereload==2.6.3
markdown2==2.4.8
MarkupSafe==2.1.2
npm==0.1.1
npx==0.1.1
numpy==1.24.2
optional-django==0.1.0
Pillow==9.4.0
Pygments==2.14.0
python-dotenv==0.21.0
pytz==2022.6
react==4.3.0
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
tornado==6.2
tzdata==2022.6
urllib3==1.26.15
vite==1.5.2
